hw1/requirements.txt
Pillow==3.1.2